+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Added support for multiple deployment targets with xcode 14
 - Added `supportedPlatforms` for target
 - Added `platformFilters` for sources and dependecies
-- Added `inferPlatformFiltersByPath`, a convenience filer for sources
+- Added `inferPlatformFiltersByPath`, a convenience filter for sources
 
 ### Added